// scripts/init/API.ts
export type CreateEmployeeInput = {
  username: string,
  companyID: string,
  no: string,
  groupID: string | null | undefined,
  superiorUsername: string,
  firstName: string,
  lastName: string,
  grade: string,
  email: string,
  sub?: string | null,
  manager: EmployeeType,
  isCompanyAdmin?: boolean | null,
  isDeleted: BooleanType,
};

export enum EmployeeType {
  OTHER = "OTHER",
  NORMAL = "NORMAL",
  MANAGER = "MANAGER",
  SUPER_MANAGER = "SUPER_MANAGER",
}


export enum BooleanType {
  TRUE = "TRUE",
  FALSE = "FALSE",
}
// ...
export type CreateGroupInput = {
  id?: string | null,
  companyID: string,
  no: string,
  name: string,
  createdAt?: string | null,
};
// ...
export type CreateCompanyInput = {
  id?: string | null,
  name: string,
  startMonth: number,
  shortName?: string | null,
  url?: string | null,
  isContractEvaluation?: boolean | null,
  isContractReport?: boolean | null,
};